pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dlib-bin>=19.24.6",
    "flask>=3.1.2",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "psutil>=7.0.0",
    "pygame>=2.6.1",
    "requests>=2.32.5",
    "scipy>=1.16.1",
]
